Extract ULWeb3Provider to it's own file
